Avoid MinGW stat wrapper on Windows
stat() is a mingw posix wrapper for the native Windows _stat() and
friends. Although tempting to use for simplicity, the wrapper incurs
overhead, such as making distribution of static libraries challenging
(as the mingw version has to match exactly). Additionally some mingw
releases are buggy and generate symbol linkage to missing symbols. Avoid
the middleman and utilize the windows API directly.